+++ src/main/java/com/nosql/springboot/common/Validator.java
@@ -2,7 +2,7 @@ package com.nosql.springboot.common;
 
 public class Validator {
 
-    public static void checkIdIsNull(Long id) {
+    public static void checkIdIsNull(Integer id) {
         if (id == null) {
             throw new NullIdException("Provided id is null");
         }
